feat: US006 - Extract Magic Numbers to Constants
